Move RecommendedTracks logic to redux (WIP)
* Improve Track/RecommendedTrack type checking.
