+++ src/features/itemsSlice.ts
@@ -1,6 +1,13 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
 
-import { Track, Playlist, Album, SortedPlaylists, SearchResultsType } from '../types';
+import {
+  Track,
+  Playlist,
+  Album,
+  SortedPlaylists,
+  SearchResultsType,
+  RecommendedTrack,
+} from '../types';
 import { AppThunk, RootState } from '../app/store';
 import { createRequestUrl, spotifyBaseRequest } from '../utils/RequestUtils';
 import axios from 'axios';
@@ -25,7 +32,7 @@ export interface ItemsState {
   tracks: Track[] | null;
   sortedTracks: Track[] | null;
 
-  recommendedTrackSeed: Track | null;
+  recommendedTrackSeed: RecommendedTrack | null;
   lastClickedTrack: string | null;
 }
 
@@ -64,7 +71,7 @@ const itemsSlice = createSlice({
     setAlbums: (state, action: PayloadAction<Album[]>) => {
       state.albums = action.payload;
     },
-    setRecommendedTrack: (state, action: PayloadAction<Track>) => {
+    setRecommendedTrack: (state, action: PayloadAction<RecommendedTrack>) => {
       state.recommendedTrackSeed = action.payload;
     },
     setLastClickedTrack: (state, action: PayloadAction<string>) => {
@@ -561,6 +568,170 @@ export const getTracks = (currentPlaylist: Playlist): AppThunk => {
     } catch (err) {
       // if (err.response?.status === 401) handleAuthError();
       console.log(err.message);
+    }
+  };
+};
+
+export const getRecommendedTracks = (recommendedTrack: RecommendedTrack): AppThunk => {
+  return async (dispatch, getState) => {
+    await dispatch(setSortedTracks(null));
+
+    const { matchRecsToSeedTrackKey, seedAttributes } = getState().controlsSlice;
+    const spotifyToken = getState().settingsSlice.spotifyToken;
+
+    const generateRecommendedUrl = (currentUrl: string, limit = 10) => {
+      let newUrl = currentUrl + `&limit=${limit}`;
+
+      // // other available api seeds
+      // if (artistSeed) url += `&seed_artists=${ artistSeed.map(artist => artist.id).join(',') }`;
+      // if (trackSeed) url += `&seed_tracks=${ trackSeed.map(track => track.id).join(',') }`;
+      // if (mode) url += `&target_mode=${mode}`
+
+      Object.keys(seedAttributes).forEach((param) => {
+        if (seedAttributes[param].value !== '') {
+          if (param === 'genre') {
+            newUrl += `&seed_genres=${seedAttributes[param].value}`;
+          } else if (param === 'duration') {
+            // Convert seconds to ms + ts casting weirdness
+            const durationInMs = parseInt(seedAttributes[param].value || '1') * 1000;
+            newUrl += `&${seedAttributes[param].maxOrMinFilter}_${param}=${durationInMs}`;
+          } else {
+            newUrl += `&${seedAttributes[param].maxOrMin}_${param}=${seedAttributes[param].value}`;
+          }
+        }
+      });
+
+      console.log("newUrl", newUrl)
+
+      return newUrl;
+    };
+
+    try {
+      const getTracksFromSpotify = async (url: string, limit: number) => {
+        const recommendedTracksUrls = generateRecommendedUrl(url, limit);
+        const tracks = await spotifyBaseRequest(spotifyToken).get(
+          recommendedTracksUrls
+        );
+
+        return tracks.data.tracks;
+      };
+
+      let tracklist;
+      if (matchRecsToSeedTrackKey) {
+        // match key
+        const url1 =
+          `https://api.spotify.com/v1/recommendations?market=AU&seed_tracks=${recommendedTrack.id}` +
+          // conditional here to account for any missing track attributes
+          `${
+            recommendedTrack.key
+              ? `&target_key=${recommendedTrack.key}`
+              : ''
+          }` +
+          `${
+            recommendedTrack.mode
+              ? `&target_mode=${recommendedTrack.mode}`
+              : ''
+          }`;
+
+        const tracks1 = await getTracksFromSpotify(url1, 25);
+
+        // minor/major alternative scale ---> if you request similar tracks for a
+        // minor scale, we also get tracks from the major scale (that share the same notes)
+        const url2 =
+          `https://api.spotify.com/v1/recommendations?market=AU&seed_tracks=${recommendedTrack.id}` +
+          `${
+            recommendedTrack.parsedKeys[2][0]
+              ? `&target_key=${recommendedTrack.parsedKeys[2][0]}`
+              : ''
+          }` +
+          `${
+            recommendedTrack.parsedKeys[2][1]
+              ? `&target_mode=${recommendedTrack.parsedKeys[2][1]}`
+              : ''
+          }`;
+        const tracks2 = await getTracksFromSpotify(url2, 15);
+
+        tracklist = tracks1.concat(tracks2);
+      } else {
+        // Recommendations without key param
+        const url = `https://api.spotify.com/v1/recommendations?market=AU&seed_tracks=${recommendedTrack.id}`;
+
+        tracklist = await getTracksFromSpotify(url, 40);
+      }
+
+      // remove any null items
+      tracklist = tracklist.filter(Boolean);
+
+      // remove duplicates (result of multiple calls)
+      tracklist = tracklist.reduce((accumulator, current) => {
+        if (!accumulator.find((el) => el.id === current.id)) {
+          accumulator.push(current);
+        }
+
+        return accumulator;
+      }, []);
+
+      const trackIds = tracklist.map((track) => track.id);
+      const artistIds = tracklist.map((track) => track.artists[0].id);
+
+      const trackFeaturesResponse: any = await spotifyBaseRequest(spotifyToken).get(
+        `audio-features/?ids=${trackIds.join(',')}`
+      );
+      const artistFeaturesResponse: any = await spotifyBaseRequest(spotifyToken).get(
+        `artists?ids=${artistIds.join(',')}`
+      );
+
+      const trackFeatures = [...trackFeaturesResponse.data.audio_features];
+      const artistFeatures = [...artistFeaturesResponse.data.artists];
+
+      const splicedTracks = tracklist
+        .filter((_, index) => trackFeatures[index] != null)
+        .map((item, index) => {
+          return {
+            name: item.name,
+            artists:
+              item.artists.length > 1
+                ? [item.artists[0].name, item.artists[1].name]
+                : [item.artists[0].name],
+            id: item.id && item.id,
+            tempo:
+              trackFeatures[index] != null
+                ? Math.round(trackFeatures[index].tempo)
+                : '',
+            key: trackFeatures[index] != null ? trackFeatures[index].key : '',
+            mode:
+              trackFeatures[index] != null ? parseInt(trackFeatures[index].mode) : '',
+            energy:
+              trackFeatures[index] != null
+                ? Math.round(trackFeatures[index].energy.toFixed(2) * 100) / 100
+                : '',
+            danceability:
+              trackFeatures[index] != null ? trackFeatures[index].danceability : '',
+            acousticness:
+              trackFeatures[index] != null ? trackFeatures[index].acousticness : '',
+            liveness: trackFeatures[index] != null ? trackFeatures[index].liveness : '',
+            loudness: trackFeatures[index] != null ? trackFeatures[index].loudness : '',
+            speechiness:
+              trackFeatures[index] != null ? trackFeatures[index].speechiness : '',
+            valence: trackFeatures[index] != null ? trackFeatures[index].valence : '',
+
+            duration:
+              item.duration_ms != null
+                ? millisToMinutesAndSeconds(item.duration_ms)
+                : '',
+            track_popularity: item.popularity != null ? item.popularity : '',
+            artist_genres:
+              artistFeatures[index] != null ? artistFeatures[index].genres : '',
+            album: item.album.name ? item.album.name : '',
+            release_date: item.album.release_date ? item.album.release_date : '',
+          };
+        });
+
+      dispatch(setTracks(splicedTracks));
+      dispatch(setSortedTracks(splicedTracks));
+    } catch (err) {
+      console.log(err.message);
+      // if (err.response?.status === 401) handleAuthError();
     }
   };
 };
